be-flashcards/requirements.txt
anyio==4.0.0
certifi==2023.7.22
click==8.1.7
colorama==0.4.6
dnslib==0.9.23
dnspython==2.4.2
exceptiongroup==1.1.3
fastapi==0.99.1
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
iniconfig==2.0.0
mongomock==4.1.2
mongomock-motor==0.0.21
motor==3.3.1
packaging==23.2
pluggy==1.3.0
pydantic==1.10.13
pymongo==4.5.0
pytest==7.4.3
pytest-asyncio==0.21.1
python-dateutil==2.8.2
python-decouple==3.8
PythonDNS==0.1
pytz==2023.3.post1
sentinels==1.0.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tomli==2.0.1
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
uvicorn==0.23.2
zope.interface==6.1
